Update translates page
