--- packages/ui/src/admin/DataTables/TranslateDataTable.tsx
@@ -1,80 +1,0 @@
-import { FC, useEffect, useState } from 'react'
-
-import { useUpdateEffect } from '@chakra-ui/react'
-import { useRouter } from 'next/router'
-
-import { useStrapiRequest } from '@wsvvrijheid/services'
-import {
-  Activity,
-  ApprovalStatus,
-  Sort,
-  StrapiCollectionEndpoint,
-  StrapiModel,
-} from '@wsvvrijheid/types'
-
-import { DataTable } from './DataTable'
-import { TranslateDataTableProps } from './types'
-import { WTableProps } from '../../components'
-import { useColumns } from '../../data'
-
-export const TranslateDataTable: FC<TranslateDataTableProps> = ({
-  searchTerm,
-}) => {
-  const [currentPage, setCurrentPage] = useState<number>(1)
-  const [sort, setSort] = useState<Sort>()
-
-  const { query, locale, push } = useRouter()
-  const status = query.status as ApprovalStatus
-  const slug = query.slug as Partial<StrapiCollectionEndpoint>
-
-  const columns = useColumns()
-
-  useEffect(() => setCurrentPage(1), [status])
-
-  useUpdateEffect(() => {
-    dataQuery.refetch()
-  }, [locale, searchTerm, sort, status])
-
-  const dataQuery = useStrapiRequest<Activity>({
-    endpoint: slug,
-    page: currentPage || 1,
-    pageSize: 10,
-    filters: {
-      ...(searchTerm && {
-        $or: [
-          { title: { $containsi: searchTerm } },
-          { description: { $containsi: searchTerm } },
-        ],
-      }),
-      approvalStatus: { $eq: 'pending' },
-    },
-    sort,
-    locale,
-    includeDrafts: true,
-  })
-
-  const items = dataQuery?.data?.data
-  const totalCount = dataQuery?.data?.meta?.pagination?.pageCount || 0
-
-  const handleClick = (index: number, id: number) => {
-    push({ query: { ...query, id } })
-  }
-
-  const mappedActivities =
-    items?.map(item => ({
-      ...item,
-      translates: item.localizations?.map(l => l.locale),
-    })) || []
-
-  return (
-    <DataTable<StrapiModel>
-      columns={columns[slug] as WTableProps<StrapiModel>['columns']}
-      data={mappedActivities}
-      totalCount={totalCount}
-      currentPage={currentPage}
-      setCurrentPage={setCurrentPage}
-      onSort={setSort}
-      onClickRow={handleClick}
-    />
-  )
-}
