+++ packages/ui/src/admin/DataTables/index.ts
@@ -1,3 +1,2 @@
 export * from './ArtsTable'
 export * from './DataTable'
-export * from './TranslateDataTable'
