+++ packages/ui/src/admin/DataTables/types.ts
@@ -7,7 +7,3 @@ export type DataTableProps<T extends StrapiModel> = {
   currentPage: number
   setCurrentPage: (page: number) => void
 } & Pick<WTableProps<T>, 'data' | 'columns' | 'onClickRow' | 'onSort'>
-
-export type TranslateDataTableProps = {
-  searchTerm?: string
-}
