+++ packages/ui/src/admin/FormLocaleSwitcher/FormLocaleSwitcher.tsx
@@ -2,20 +2,23 @@ import { Button, ButtonGroup } from '@chakra-ui/react'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 
-import { StrapiTranslatableModel } from '@wsvvrijheid/types'
+import { StrapiLocale, StrapiTranslatableModel } from '@wsvvrijheid/types'
 
 export const FormLocaleSwitcher = <T extends StrapiTranslatableModel>({
   models,
+  currentLocale,
 }: {
   models: T[]
-  slug?: string
+  currentLocale: StrapiLocale
 }) => {
   const router = useRouter()
   const slug = router.query.slug as string
 
   return (
     <ButtonGroup>
-      {models?.map(model => {
+      {router.locales?.map(locale => {
+        const model = models.find(model => model.locale === locale) as T
+
         const href =
           slug && router.pathname !== '/translates'
             ? `/${slug}/${model.id}`
@@ -30,6 +33,7 @@ export const FormLocaleSwitcher = <T extends StrapiTranslatableModel>({
             key={model.id}
             textTransform={'uppercase'}
             href={href}
+            variant={model.locale === currentLocale ? 'solid' : 'outline'}
           >
             {model.locale}
           </Button>
